Death transition
The fade to black is accomplished by having a UI image that covers the entire screen that is completely black. It is animated such that triggering the animation causes the opacity to goes from zero (where it is normally) to one and back down to zero. Instead of animating the opacity of the 'color' on the 'image', Brackeys recommended adding a canvas group component and animating its alpha. By doing this, it should allow us to fade text and other screen elements if we want.

I added the trigger for the transition into the 'player' class. When the player loses all hp, the fadeToBlack() gets called from the crossfade class.

Quite possible we want to adjust timing. Currently just have 1 second to black and 1 second to clear again.

The script also starts the scene by going from black to the game. I think this looks good, but we can turn it off if needed.