AudioController
Class AudioController has a function changed changeTrack which takes as arguments 2 audioSources. The first audiosource fades to zero volume over 5 seconds (can easily be adjusted on a public variable in AudioController). The second Audiosource starts playing at zero volume and gradually fades up to full volume over 5 seconds. (With fades in both cases being done by class FadeAudioSource.)

Depending on how the rest of the code and scene are setup, it may be possible to determine the currently playing track with code and reduce the function down to one argument. However, I think I would need to see more of the game in action before that is possible.