Fixing fade to black
Steps to get the crossfade to work:
Added a UI image to canvas. Assign image color to straight black. Add a canvas group component and set alpha to zero (so you can see the game while building it :) )
Add the crossfade Animator (in the animator folder) to the black UI image.
Add the crossfade script to canvas (was already there)
Assign "crossfade animator" in canvas by dragging in the black UI image.

There was also an issue in player code. The code had it set to fade to black when you shot, not when you died. The fade to black is now after death.